requirements.txt
rfc3339==5
google-api-python-client
django-model-utils==2.0.3
tzlocal==1.1.1
celery==3.1.11
six==1.10
uritemplate==3.0.0
simplejson==3.10
